backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = ""
authors = [
    {name = "ILikeToasts4",email = "jerome.turgeon-nozaki.1@ens.etsmtl.ca"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "spotipy (>=2.25.1,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "requests (>=2.32.4,<3.0.0)",
    "lyricsgenius (>=3.6.5,<4.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-ollama (>=0.3.6,<0.4.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jerome"]
package-mode = false
[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"